refactor filesystem plugins directory layout
