+++ tensorflow_io/core/filesystems/az/az_filesystem.cc
@@ -34,7 +34,7 @@ limitations under the License.
 #include "storage_errno.h"
 #include "tensorflow/c/logging.h"
 #include "tensorflow/c/tf_status.h"
-#include "tensorflow_io/core/plugins/file_system_plugins.h"
+#include "tensorflow_io/core/filesystems/filesystem_plugins.h"
 
 namespace tensorflow {
 namespace io {
