+++ tensorflow_io/core/filesystems/http/http_filesystem.cc
@@ -24,7 +24,7 @@ limitations under the License.
 #include "absl/synchronization/mutex.h"
 #include "tensorflow/c/logging.h"
 #include "tensorflow/c/tf_status.h"
-#include "tensorflow_io/core/plugins/file_system_plugins.h"
+#include "tensorflow_io/core/filesystems/filesystem_plugins.h"
 
 namespace tensorflow {
 namespace io {
