+++ tensorflow_io/core/filesystems/s3/s3_filesystem.cc
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow_io/core/plugins/s3/s3_filesystem.h"
+#include "tensorflow_io/core/filesystems/s3/s3_filesystem.h"
 
 #include <aws/core/client/AsyncCallerContext.h>
 #include <aws/core/config/AWSProfileConfigLoader.h>
@@ -38,8 +38,8 @@ limitations under the License.
 #include "absl/strings/str_cat.h"
 #include "tensorflow/c/logging.h"
 #include "tensorflow/c/tf_status.h"
-#include "tensorflow_io/core/plugins/file_system_plugins.h"
-#include "tensorflow_io/core/plugins/s3/aws_logging.h"
+#include "tensorflow_io/core/filesystems/filesystem_plugins.h"
+#include "tensorflow_io/core/filesystems/s3/aws_logging.h"
 
 namespace tensorflow {
 namespace io {
