+++ tensorflow_io/core/filesystems/s3/s3_filesystem.h
@@ -12,8 +12,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#ifndef TENSORFLOW_C_EXPERIMENTAL_FILESYSTEM_PLUGINS_S3_S3_FILESYSTEM_H_
-#define TENSORFLOW_C_EXPERIMENTAL_FILESYSTEM_PLUGINS_S3_S3_FILESYSTEM_H_
+#ifndef TENSORFLOW_IO_CORE_FILESYSTEMS_S3_S3_FILESYSTEM_H_
+#define TENSORFLOW_IO_CORE_FILESYSTEMS_S3_S3_FILESYSTEM_H_
 
 #include <aws/core/Aws.h>
 #include <aws/core/utils/StringUtils.h>
@@ -106,4 +106,4 @@ void RenameFile(const TF_Filesystem* filesystem, const char* src,
 }  // namespace io
 }  // namespace tensorflow
 
-#endif  // TENSORFLOW_C_EXPERIMENTAL_FILESYSTEM_PLUGINS_S3_S3_FILESYSTEM_H_
+#endif  // TENSORFLOW_IO_CORE_FILESYSTEMS_S3_S3_FILESYSTEM_H_
